Refactor account/auth related classes to their own package
