+++ app/src/main/java/dev/ridill/rivo/account/data/repository/AuthRepositoryImpl.kt
@@ -1,21 +1,21 @@
-package dev.ridill.rivo.settings.data.repository
+package dev.ridill.rivo.account.data.repository
 
 import android.app.PendingIntent
 import android.content.Intent
 import dev.ridill.rivo.R
 import dev.ridill.rivo.core.data.util.tryNetworkCall
-import dev.ridill.rivo.core.domain.model.AuthState
+import dev.ridill.rivo.account.domain.model.AuthState
 import dev.ridill.rivo.core.domain.model.DataError
 import dev.ridill.rivo.core.domain.model.Result
-import dev.ridill.rivo.core.domain.model.UserAccount
-import dev.ridill.rivo.core.domain.service.AccessTokenService
-import dev.ridill.rivo.core.domain.service.AuthService
-import dev.ridill.rivo.core.ui.authentication.AuthorizationFailedThrowable
-import dev.ridill.rivo.core.ui.authentication.AuthorizationNeedsResolutionThrowable
-import dev.ridill.rivo.core.ui.authentication.AuthorizationService
-import dev.ridill.rivo.core.ui.authentication.CredentialService
+import dev.ridill.rivo.account.domain.model.UserAccount
+import dev.ridill.rivo.account.domain.service.AccessTokenService
+import dev.ridill.rivo.account.domain.service.AuthService
+import dev.ridill.rivo.account.presentation.AuthorizationFailedThrowable
+import dev.ridill.rivo.account.presentation.AuthorizationNeedsResolutionThrowable
+import dev.ridill.rivo.account.presentation.AuthorizationService
+import dev.ridill.rivo.account.presentation.CredentialService
 import dev.ridill.rivo.core.ui.util.UiText
-import dev.ridill.rivo.settings.domain.repositoty.AuthRepository
+import dev.ridill.rivo.account.domain.repository.AuthRepository
 import kotlinx.coroutines.CancellationException
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.Flow
