+++ app/src/main/java/dev/ridill/rivo/account/domain/model/AuthState.kt
@@ -1,0 +1,6 @@
+package dev.ridill.rivo.account.domain.model
+
+sealed interface AuthState {
+    data object UnAuthenticated : AuthState
+    data class Authenticated(val account: UserAccount) : AuthState
+}
