+++ app/src/main/java/dev/ridill/rivo/account/domain/model/UserAccount.kt
@@ -1,0 +1,7 @@
+package dev.ridill.rivo.account.domain.model
+
+data class UserAccount(
+    val email: String,
+    val displayName: String,
+    val photoUrl: String
+)
