+++ app/src/main/java/dev/ridill/rivo/account/domain/service/AccessTokenSharedPrefService.kt
@@ -1,4 +1,4 @@
-package dev.ridill.rivo.core.domain.service
+package dev.ridill.rivo.account.domain.service
 
 import android.content.SharedPreferences
 import kotlinx.coroutines.Dispatchers
