+++ app/src/main/java/dev/ridill/rivo/account/presentation/DefaultCredentialService.kt
@@ -1,4 +1,4 @@
-package dev.ridill.rivo.core.ui.authentication
+package dev.ridill.rivo.account.presentation
 
 import android.content.Context
 import androidx.credentials.ClearCredentialStateRequest
