--- app/src/main/java/dev/ridill/rivo/core/domain/model/UserAccount.kt
@@ -1,12 +1,0 @@
-package dev.ridill.rivo.core.domain.model
-
-data class UserAccount(
-    val email: String,
-    val displayName: String,
-    val photoUrl: String
-)
-
-sealed interface AuthState {
-    data object UnAuthenticated : AuthState
-    data class Authenticated(val account: UserAccount) : AuthState
-}
