+++ app/src/main/java/dev/ridill/rivo/core/ui/navigation/destinations/OnboardingScreenSpec.kt
@@ -21,10 +21,10 @@ import androidx.navigation.NavBackStackEntry
 import androidx.navigation.NavHostController
 import dev.ridill.rivo.R
 import dev.ridill.rivo.application.EXTRA_RUN_CONFIG_RESTORE
-import dev.ridill.rivo.core.domain.model.AuthState
+import dev.ridill.rivo.account.domain.model.AuthState
 import dev.ridill.rivo.core.domain.util.BuildUtil
 import dev.ridill.rivo.core.domain.util.LocaleUtil
-import dev.ridill.rivo.core.ui.authentication.rememberCredentialService
+import dev.ridill.rivo.account.presentation.rememberCredentialService
 import dev.ridill.rivo.core.ui.components.CollectFlowEffect
 import dev.ridill.rivo.core.ui.components.rememberMultiplePermissionsLauncher
 import dev.ridill.rivo.core.ui.components.rememberMultiplePermissionsState
