+++ app/src/main/java/dev/ridill/rivo/onboarding/presentation/OnboardingViewModel.kt
@@ -12,20 +12,20 @@ import dagger.hilt.android.lifecycle.HiltViewModel
 import dev.ridill.rivo.R
 import dev.ridill.rivo.core.data.preferences.PreferencesManager
 import dev.ridill.rivo.core.domain.crypto.CryptoManager
-import dev.ridill.rivo.core.domain.model.AuthState
+import dev.ridill.rivo.account.domain.model.AuthState
 import dev.ridill.rivo.core.domain.model.Result
 import dev.ridill.rivo.core.domain.util.BuildUtil
 import dev.ridill.rivo.core.domain.util.EventBus
 import dev.ridill.rivo.core.domain.util.Zero
 import dev.ridill.rivo.core.domain.util.logI
-import dev.ridill.rivo.core.ui.authentication.AuthorizationService
-import dev.ridill.rivo.core.ui.authentication.CredentialService
+import dev.ridill.rivo.account.presentation.AuthorizationService
+import dev.ridill.rivo.account.presentation.CredentialService
 import dev.ridill.rivo.core.ui.util.UiText
 import dev.ridill.rivo.onboarding.domain.model.DataRestoreState
 import dev.ridill.rivo.onboarding.domain.model.OnboardingPage
 import dev.ridill.rivo.settings.domain.backup.BackupWorkManager
 import dev.ridill.rivo.settings.domain.modal.BackupDetails
-import dev.ridill.rivo.settings.domain.repositoty.AuthRepository
+import dev.ridill.rivo.account.domain.repository.AuthRepository
 import dev.ridill.rivo.settings.domain.repositoty.BackupRepository
 import dev.ridill.rivo.settings.domain.repositoty.SettingsRepository
 import kotlinx.coroutines.delay
