+++ app/src/main/java/dev/ridill/rivo/settings/presentation/settings/SettingsScreen.kt
@@ -46,7 +46,7 @@ import androidx.paging.compose.itemContentType
 import androidx.paging.compose.itemKey
 import dev.ridill.rivo.BuildConfig
 import dev.ridill.rivo.R
-import dev.ridill.rivo.core.domain.model.AuthState
+import dev.ridill.rivo.account.domain.model.AuthState
 import dev.ridill.rivo.core.domain.util.BuildUtil
 import dev.ridill.rivo.core.domain.util.Zero
 import dev.ridill.rivo.core.domain.util.tryOrNull
